+++ frontend/src/components/Calendar/index.jsx
@@ -5,10 +5,14 @@ import Month from '../Month';
 
 export default function Calendar (){
 
+    let today = new Date();
    
     return (
-        <div id="calendars">
-            <Month />
+        <div id="calendars" className='flex flex-row flex-wrap justify-center'>
+            <Month date={today} />
+            <Month date={new Date(today.getFullYear(), today.getMonth()+1, 1)} />
+            <Month date={new Date(today.getFullYear(), today.getMonth()+2, 1)} />
+            <Month date={new Date(today.getFullYear(), today.getMonth()+3, 1)} />
         </div>
     )
 }
