+++ frontend/src/components/Week/index.jsx
@@ -1,0 +1,15 @@
+import Day from '../Day';
+
+export default function Week(){
+    return (
+        <div className='flex flex-row'>
+            <Day />
+            <Day />
+            <Day />
+            <Day />
+            <Day />
+            <Day />
+            <Day />
+        </div>
+    )
+}
